+++ frontend/src/components/Header.jsx
@@ -1,9 +1,7 @@
-import React from 'react'
-
 export const Header = () => {
   return (
     <header>
-        <h1>Log in to reach the secret page</h1>
+      <h1>Log in to reach the secret page</h1>
     </header>
-  )
-}
+  );
+};
